Fix after review: * Remove unneeded whitespace changes * Extract code related with managed delegate to separate header * Remove unneded files and defines from cmake files
