+++ src/coreclr/hosts/unixcoreruncommon/CMakeLists.txt
@@ -2,10 +2,6 @@ project(unixcoreruncommon)
 
 add_compile_options(-fPIC)
 
-if(FEATURE_GDBJIT)    
-   add_definitions(-DFEATURE_GDBJIT)
-endif(FEATURE_GDBJIT)
-
 _add_library(unixcoreruncommon
     STATIC
     coreruncommon.cpp
